Separate formula classes
