+++ client/Rules/Formulas/AbilityReference.ts
@@ -1,0 +1,45 @@
+import { AbilityScores, StatBlock } from "../../../common/StatBlock";
+import { IRules } from "../Rules";
+import { FormulaTerm } from "./FormulaTerm";
+
+export class AbilityReference implements FormulaTerm {
+  private GetModifier: (score: number) => number; // TODO: where does this get injected?
+  private readonly Key: keyof AbilityScores;
+  private static KeyForPattern(pattern: string): keyof AbilityScores {
+    switch (pattern) {
+      case "STR":
+        return "Str";
+      case "DEX":
+        return "Dex";
+      case "CON":
+        return "Con";
+      case "INT":
+        return "Int";
+      case "WIS":
+        return "Wis";
+      case "CHA":
+        return "Cha";
+      default:
+        throw `Illegal ability score name ${pattern}`;
+    }
+  }
+  public readonly HasStaticResult = true;
+  public readonly RequiresStats = true;
+  public Evaluate(stats?: StatBlock) {
+    if (!stats) {
+      throw `Ability reference can't be calculated without StatBlock context!`;
+    }
+    return this.GetModifier(stats.Abilities[this.Key]);
+  }
+  public EvaluateStatic = this.Evaluate;
+  public FormulaString(): string {
+    return `${this.Key}`;
+  }
+  public static readonly Pattern = /\[(STR|DEX|CON|INT|WIS|CHA)\]/;
+  public static TestPattern = /\[(?:STR|DEX|CON|INT|WIS|CHA)\]/;
+  constructor(match: string, rules: IRules) {
+    const result = AbilityReference.Pattern.exec(match);
+    this.Key = AbilityReference.KeyForPattern(result[1]);
+    this.GetModifier = rules.GetModifierFromScore;
+  }
+}
